Add sync mode test coverage for sub account
