+++ test/test-cases/index.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const apiSyncModeSqliteTestCases = require(
+  './api-sync-mode-sqlite-test-cases'
+)
+
+module.exports = {
+  apiSyncModeSqliteTestCases
+}
